// src/app/enum/dialog.type.ts
export enum DialogType {
  USER_CREATION = 'USER_CREATION',
  USER_DELETION = 'USER_DELETION',
  USER_UPDATE = 'USER_UPDATE',
  USER_INFORMATION = 'USER_INFORMATION',
  USER_ALERT = 'USER_ALERT',

  CUSTOMER_CREATION = 'CUSTOMER_CREATION',
  CUSTOMER_DELETION = 'CUSTOMER_DELETION',
  CUSTOMER_UPDATE = 'CUSTOMER_UPDATE',
  CUSTOMER_INFORMATION = 'CUSTOMER_INFORMATION',
  CUSTOMER_ALERT = 'CUSTOMER_ALERT'
}